Fake_Face_detection/factor/FaceX-Zoo/data_processor/test_dataset.py
import os
from PIL import Image

class TestDataset:
    def __init__(self, image_folder):
        self.image_folder = image_folder
        self.image_paths = [os.path.join(image_folder, img) for img in os.listdir(image_folder)]

    def __len__(self):
        return len(self.image_paths)

    def __getitem__(self, idx):
        img_path = self.image_paths[idx]
        image = Image.open(img_path).convert('RGB')
        return image, img_path
