requirements.txt
flask
requests
pymongo
werkzeug
numpy
torch
torchvision
Pillow
urllib3
PyYAML
faiss-cpu
opencv-python-headless
tqdm
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.3
black==23.3.0
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
distro==1.9.0
exceptiongroup==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.12.2
google-search-results==2.4.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==6.7.0
jieba3k==0.35.1
joblib==1.3.2
lxml==5.2.0
lxml_html_clean==0.1.0
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.8.1
openai==1.14.3
packaging==24.0
pathspec==0.11.2
Pillow==9.5.0
platformdirs==4.0.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
serpapi==0.1.5
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
tinysegmenter==0.3
tldextract==4.0.0
tomli==2.0.1
tqdm==4.66.2
typed-ast==1.5.5
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
Flask
openai==0.28
beautifulsoup4
Flask==3.0.3
ftfy==6.2.3
ipython==8.12.3
numpy==2.1.1
opencv_python==4.9.0.80
Pillow==10.4.0
regex==2024.7.24
regex==2024.5.15
scikit_learn==1.4.1.post1
scipy==1.14.1
setuptools==69.5.1
skimage==0.0
tensorboardX==2.6.2.2
torch==2.3.1+cu118
torchvision==0.18.1
tqdm==4.66.4
util==0.0.1
Werkzeug==3.0.4
